GADE Update
This commit includes audio that can be heard for both the slash screen and the game space. More assets are included for preparation for obstacles, environment scene. Player can jump and a cube is placed as a obstacle which will help for obstacle spawning and colliding with the obstacle. the death score and pause menus function  well and display the Player's score. To conclude all errors are fixed.